% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_panel_data.R
\name{generate_panel_data}
\alias{generate_panel_data}
\title{Generate Panel Data}
\usage{
generate_panel_data(
  n_units = 1000,
  n_time = 10,
  e_mult = 3,
  e_share = 0.1,
  share_missing = 0,
  mean_u = 0,
  sigma_u = 1/3,
  rho_u = 0.5,
  mean_t = 0,
  sigma_t = 1/3,
  rho_t = 0.5,
  sigma_e = 1 - sigma_u - sigma_t,
  beta = 1,
  mu_x = 0,
  sigma_x = 1,
  n_z = 1,
  beta_z = 0,
  mu_z = 0,
  sigma_z = 1,
  seed = NULL
)
}
\arguments{
\item{n_units}{Integer. Number of cross-sectional units. Default is 1000.}

\item{n_time}{Integer. Number of time periods. Default is 10.}

\item{e_mult}{Numeric. Multiplier to set the kurtosis fo the composite error.
\code{e_share} of the composite error will be multiplied by \code{e_mult}.
Default is 3, resulting in mild kurtosis of the composite error.}

\item{e_share}{Numeric. Share of observations affected by the error
multiplier. Default is 0.1.}

\item{share_missing}{Numeric. Percentage of missing data to introduce into
the panel. Default is 0 (no missing data, balanced panel). Needs to be
between 0 and smaller than 1.}

\item{mean_u}{Numeric. Mean of the unit-level random effects. Default is 0.}

\item{sigma_u}{Numeric. Standard deviation of the unit-level random effects.
Default is 1/3.}

\item{rho_u}{Numeric. Unit-specific autocorrelation coefficient for the
error term. Default is 0.5.}

\item{mean_t}{Numeric. Mean of the time-level random effects. Default is 0.}

\item{sigma_t}{Numeric. Standard deviation of the time-level random effects.
Default is 1/3.}

\item{rho_t}{Numeric. Time-specific autocorrelation coefficient for the
error term. Default is 0.5.}

\item{sigma_e}{Numeric. Standard deviation of the idiosyncratic noise
component. Default is 1 - sigma_u - sigma_t.}

\item{beta}{Numeric. Coefficient for the independent variable x. Default is 1.}

\item{mu_x}{Numeric. Mean of the independent variable x. Default is 0.}

\item{sigma_x}{Numeric. Standard deviation of the independent variable x.
Default is 1.}

\item{n_z}{Integer. Number of control variables z to include. Default is 1.}

\item{beta_z}{Numeric vector. Coefficients for the control variable(s) z.
Default is 0 (no effect). Needs to be of length \code{n_z}.}

\item{mu_z}{Numeric vector. Mean of the control variable(s) z. Default is 0.
Needs to be of length \code{n_z}.}

\item{sigma_z}{Numeric vector. Standard deviation of the control
variable(s) z. Default is 1. Needs to be of length \code{n_z}.}

\item{seed}{Integer. Random seed for reproducibility.}
}
\value{
A data frame containing the simulated panel data.
}
\description{
Generates a data frame containing simulated panel data for testing the
\code{ferols()} function. The data contains observations with unit and
time dimensions and two-way clustered composite errors drawn from a mixed
distribution.
}
\examples{
df <- generate_panel_data(n_units = 50, n_time = 10, seed = 42)
head(df)
ferols(y ~ x + z | i + t, vcov = ~i, data = df)
fixest::feols(y ~ x + z | i + t, vcov = ~i, data = df)

}
