% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferols.R
\name{vcov.ferols}
\alias{vcov.ferols}
\title{Variance–covariance matrix for \code{ferols} objects}
\usage{
\method{vcov}{ferols}(object, vcov = NULL, cluster = NULL, se = NULL, ssc = NULL, ...)
}
\arguments{
\item{object}{A \code{ferols} object.}

\item{vcov}{Variance–covariance specification. See
\code{\link[fixest:vcov.fixest]{fixest::vcov.fixest()}} for details. Defaults to heteroskedasticity-robust
(White) standard errors.}

\item{cluster}{Deprecated. Use \code{vcov} instead.}

\item{se}{Deprecated. Use \code{vcov} instead.}

\item{ssc}{Small-sample correction. Passed on to
\code{"fixest::vocov.fixest"}.}

\item{...}{Additional arguments (passed on to \code{"fixest::vocov.fixest"}).}
}
\value{
A variance–covariance matrix with coefficient names as row and
column names. The matrix carries a \code{"type"} attribute describing
the estimator.
}
\description{
Computes variance–covariance matrices for estimates generated by
\link{ferols}. For "iid" standard errors, the method mirrors
the approach of \code{\link[MASS:rlm]{MASS::rlm()}}. All other error types are
relegated to \code{\link[fixest:vcov.fixest]{fixest::vcov.fixest()}} after the bread
part of the sandwich estimator is adjusted as discussed in Appendix A.2 of
\href{https://doi.org/10.1016/j.jacceco.2026.101895}{Gassen and Veenman (2026)}.
This has been verified to yield consistent standard errors for (multiway)
clustered SEs but should also work for other SEs (e.g., Newey-West,
Discroll-Kray, and Conley) that follow the sandwich estimator structure.
Use with care for these and feel free to open an issue on GitHub
if you sense any inconsistencies.
}
\details{
This function is experimental and implements only a subset of the
variance estimators available in \code{fixest}. Several panel robust
estimators and user-supplied functions and covariance matrices are not
supported.
}
