% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferols.R
\name{ferols}
\alias{ferols}
\title{Fixed-effects robust IRLS M regression (Huber loss)}
\usage{
ferols(
  fml,
  data,
  family = "huber",
  efficiency = 0.95,
  k = NULL,
  scale_est = NULL,
  scale = NULL,
  scale_update = FALSE,
  tol = 1e-10,
  max_iter = 200,
  adj_rlm = FALSE,
  rm_singletons = TRUE,
  vcov = NULL,
  ssc = NULL,
  cluster = NULL,
  se = NULL,
  ...
)
}
\arguments{
\item{fml}{A fixest formula of the form \code{y ~ x1 + x2 | fe1 + fe2}.}

\item{data}{A data.frame.}

\item{family}{Robust loss family. Currently only \code{"huber"}.}

\item{efficiency}{Target normal-efficiency in (0.68, 1). Default is 0.95.}

\item{k}{Huber Loss function parameter. If \code{NULL} (the default), it will be
determined by the \code{efficiency} parameter. If set, it will over-rule the
\code{efficiency} parameter.}

\item{scale_est}{Algorithm to estimate the residuals that are used to
estimate the (inital) location and scale. Defaults to \code{"lad_rq"} when no
fixed effects are present. Otherwise, it defaults to \code{"lad_mm_rsc"}.
\itemize{
\item \code{"ols"}: Uses \code{fixest:feols()} with absorbed fixed effects to derive
the initial residuals.
\item \code{"lad_rq"}: Uses absolute residuals from a normal median/LAD regression
(calling \code{quantreg::rq()}) to estimate the scale. Can only be used when
no fixed effects are present.
\item \code{"lad_mm_ms"}: Uses an method of moments algorithm for the median
quantile regression introduced by Machado and Santos Silva
(2019, https://doi.org/10.1016/j.jeconom.2019.04.009).
This should yield similar scale and beta estimates as the approaches of
the Stata packages \code{robreg} (https://github.com/benjann/robreg) and
\code{robhdfe} (https://github.com/dveenman/robhdfe).
\item \code{"lad_mm_rsc"}: Method of moments algorithm as extended by
Rios-Avila, Siles, and Canavire-Bacarreza
(2024, https://docs.iza.org/dp17262.pdf). This algorithm is
faster as it absorbs multi-dimensional fixed effects. Its scale and beta
estimates should be virtually identical (within numerical precision) to
the ones generated by \code{"lad_mm_ms"}.
}}

\item{scale}{A positive numerical value to override the estimated scale.}

\item{scale_update}{Should the scale estimate be updated in every IRLS step?
Defaults to \code{FALSE} like Stata's \code{robreg}. Set to \code{TRUE} to get closer to
\code{MASS::rlm()} behavior.}

\item{tol}{Convergence tolerance on relative coefficient change.}

\item{max_iter}{Maximum IRLS iterations.}

\item{adj_rlm}{Whether or not the algorithm should be adjusted to be closer
ro \code{MASS::rlm()}. Setting this to \code{TRUE} implies
\itemize{
\item scale estimates to be calculated by \code{median(abs(resid_tau))/0.6745}
instead by a function that factors in the degrees of freedom, and
\item the convergence criterion to be based on residuals instead of parameter
changes.
}}

\item{rm_singletons}{Whether singletons should be removed when estimating
inintial coefficients and scale as well as during IRLS steps. Defaults to
\code{TRUE} (\code{fixest::feols()}  default). Set to \code{FALSE} to mimic the current
behavior of Stata's \code{robtwfe}.}

\item{vcov}{Variance–covariance specification. Either \code{"iid"},
\code{"hetero"}, \code{"cluster"}, \code{"twoway"} or a formula
such as \code{~ id} or  \code{cluster ~ id + time}.
Defaults to heteroskedasticity-robust (White) standard errors.}

\item{ssc}{Small-sample correction. Passed on to
\code{"fixest::vocov.fixest"}.}

\item{cluster}{(deprecated) clustering variable name(s) (string).
Note that this argument is deprecated, you should use \code{vcov} instead.}

\item{se}{(deprecated) A string ("standard", or "cluster") to indicated how
the standard errors should be calculated. Note that this argument is
deprecated, you should use \code{vcov} instead.}

\item{...}{Further arguments passed to \code{fixest::feols}. The \code{weights} and
\code{fixef.rm} arguments cannot be used as
\itemize{
\item \code{weight} is set by the IRLS process and regression weights are not
implemented yet by \code{ferols()}, and
\item \code{fixef.rm} is controlled by \code{rm_singletons}.
}}
}
\value{
An object of class \verb{c("ferols", fixest")} with an additional list
\code{robust} containing the following members:
\itemize{
\item \code{family}: The family of the loss function (currently only \code{"huber"})
\item \code{scale_est}: The function to generate the initial estimates of location
and scale. See discussion in the argument list above.
\item \code{efficiency}: the desired efficiency under normality, determining the
\code{k}tuning parameter.
\item \code{k}: The tuning parameter of the Huber loss function.
\item \code{scale}: The scale estimate for standardizing the residuals. If
\code{scale_update = FALSE} (the default), it is the outcome of the initial
scale estimation. Otherwise, it is the scale estimation resulting from
the final IRLS step.
\item \code{converged}: Whether the IRLS process converged.
\item \code{iter}: The number of IRLS runs it took for convergence or \code{max_iter} if
the algorithm did not converge.
\item \code{weights}: The weights resulting from the Huber loss weighting of the
final estimation step. These weights are informative as they identify
the 'bulk of the data' that remains influential for the robust regression
estimates.
}
}
\description{
Estimates linear models with high-dimensional fixed effects using
iteratively reweighted least squares (IRLS) for Huber M-estimation.
Fixed effects are absorbed using the approach of \code{fixest::demean}.
Standard errors can be clustered via multiple levels using a Huber
(psi/phi) sandwich estimator.
}
\details{
\code{ferols()} implements an IRLS process based on initial location and scale
estimates to derive a robust M estimate. During both stages, it uses
\code{fixest::feols} and its demeaning approach to absorb high-dimensional
fixed effects. This approach generates coefficient and standard error
estimates that are very close to either \code{MASS::rlm()} or Stata's \verb{robreg  m}.
By default, \code{ferols()} mimics the algorithm of \code{robreg}. To generate
estimates that are identical to \code{MASS::rlm()} use a call like
\code{ferols(fml, data, k = 1.345, scale_est = "ols", adj_rlm = TRUE, scale_update = TRUE, tol = 1e-4, vcov = "iid")}.

For the first-step quantile regression that is used to obtain the scale
estimate, \code{ferols()} implements the MM-QR estimation from
\href{https://www.sciencedirect.com/science/article/pii/S0304407619300648}{Machado and Santos Silva(2019)}
to combine quantile estimation with fixed effects, which
\href{https://papers.ssrn.com/sol3/papers.cfm?abstract_id=4944894}{Rios-Avila, Siles, and Canavire-Bacarreza (2024)}
show can be extended to the multidimensional fixed effects setting.

\code{ferols()} drops singleton observations by default. This implementation choice
affects not only the degrees-of-freedom adjustment in the standard error
calculation (see \href{https://scorreia.com/research/singletons.pdf}{Correia 2015}),
but in this case also ensures the scale parameter is correctly estimated with
the MM-QR estimation. Singleton observations produce regression residuals that
are equal to (or very close to) zero, causing the scale parameter to be
understated when singleton observations are retained.

By default, \code{ferols()} relays the messages and warnings from the initial
location as well as the final IRLS estimation to the user. This behavior can
be changed by setting the \code{warn} and \code{notes} arguments of \code{fixest::feols()}
}
\examples{
df <- generate_panel_data(n_units = 50, n_time = 10, seed = 42)
ferols(y ~ x + z | i + t, vcov = ~i, data = df)
fixest::feols(y ~ x + z | i + t, vcov = ~i, data = df)

}
