% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferols.R
\name{ferols}
\alias{ferols}
\title{Fixed-effects robust IRLS M regression (Huber loss)}
\usage{
ferols(
  fml,
  data,
  family = "huber",
  efficiency = 0.95,
  scale_est = "ols",
  scale = NULL,
  tol = 1e-10,
  max_iter = 200,
  vcov = NULL,
  cluster = NULL,
  ssc = NULL,
  se = NULL,
  ...
)
}
\arguments{
\item{fml}{A fixest formula of the form \code{y ~ x1 + x2 | fe1 + fe2}.}

\item{data}{A data.frame.}

\item{family}{Robust loss family. Currently only \code{"huber"}.}

\item{efficiency}{Target normal-efficiency in (0.68, 1). Default 0.95.}

\item{scale_est}{Algorithm to estimate the residuals that are used to
estimate the scale. Defaults to \code{"ols"}.
\itemize{
\item \code{"ols"}: Uses \code{fixest:feols()} with absorbed fixed effects to derive
the initial residuals.
\item \code{"lad_rq"}: Uses \code{quantreg:rq.fit()} with absorbed fixed effects
to estimate a median quantile regression (aka least absolute deviation
(LAD) regression) to derive the initial residuals.
\item \code{"lad_mm"}: Uses an method of moments algorithm for the median quantile
regression introduced by Machado and Santos Silva
(2019, https://doi.org/10.1016/j.jeconom.2019.04.009) and extended by
Rios-Avila, Siles, and Canavire-Bacarreza
(2024, https://docs.iza.org/dp17262.pdf). This should yield similar
scales as the approaches of the Stata packages \code{robreg}
(https://github.com/benjann/robreg) and \code{robtwfe}
(https://github.com/dveenman/robtwfe). This algorithm is currently
woirk-in-process.
}}

\item{scale}{A positive numerical value to override the estimated scale.}

\item{tol}{Convergence tolerance on relative coefficient change.}

\item{max_iter}{Maximum IRLS iterations.}

\item{vcov}{Variance-covariance specification. Either a string
(\code{"iid"} or \code{"cluster"}) or a formula (\code{cluster ~ var}) to indicate the
how it should be estimated.
\itemize{
\item \code{"iid"} for non-clustered Huber sandwich estimator (classical vcov,
default when no fixed effects are present).
\item \code{"cluster"} to cluster on the first fixed effect (default when fixed
effects are present).
\item \code{cluster ~ id} or simply \code{~id} for one-way clustering by the specified
variable.
All other features provided by \code{fixest} are not implemented (yet).
}}

\item{cluster}{(deprecated) A one-way clustering variable name (string).
Note that this argument is deprecated, you should use \code{vcov} instead.}

\item{ssc}{Small sample correction for constructing the covariance matrix.
The code uses a default SSC of the type (G / (G - 1)) * ((n - 1) / (n - p)).
This correction cannot be modified, so this parameter has to stay \code{NULL}
for the time being.}

\item{se}{(deprecated) A string ("standard", or "cluster") to indicated how
the standard errors should be calculated. Note that this argument is
deprecated, you should use \code{vcov} instead.}

\item{...}{Further arguments passed to \code{fixest::feols}. The \code{weights}
argument cannot be used as it is set by the IRLS process.}
}
\value{
An object of class \code{"fixest"} with an additional list
\code{robust} containing \code{k}, \code{scale}, \code{efficiency}, \code{converged}, \code{iter},
\code{weights}, and \code{phi}.
}
\description{
Estimates linear models with high-dimensional fixed effects using
iteractively reweighted least squares (IRLS) for Huber M-estimation.
Fixed effects are absorbed using \code{fixest::demean}.
Standard errors can be clustered via one dimension using a Huber (psi/phi)
sandwich estimator.
}
