% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferols.R
\name{vcov.ferols}
\alias{vcov.ferols}
\title{Variance–covariance matrix for \code{ferols} objects}
\usage{
\method{vcov}{ferols}(object, vcov = NULL, cluster = NULL, se = NULL, ssc = NULL, ...)
}
\arguments{
\item{object}{A \code{ferols} object.}

\item{vcov}{Variance–covariance specification. Either \code{"iid"},
\code{"hetero"}, \code{"cluster"}, \code{"twoway"} or a formula
such as \code{~ id} or  \code{cluster ~ id + time}.
Defaults to heteroskedasticity-robust (White) standard errors.}

\item{cluster}{Deprecated. Use \code{vcov} instead.}

\item{se}{Deprecated. Use \code{vcov} instead.}

\item{ssc}{Small-sample correction. Passed on to
\code{"fixest::vocov.fixest"}.}

\item{...}{Additional arguments (passed on to \code{"fixest::vocov.fixest"}).}
}
\value{
A variance–covariance matrix with coefficient names as row and
column names. The matrix carries a \code{"type"} attribute describing
the estimator.
}
\description{
Computes variance–covariance matrices for estimates generarted by
\code{\link{ferols}}. The interface mirrors \code{fixest::vcov.fixest},
but currently supports only Huber M-estimation sandwich estimators
with either no clustering or one-way clustering.
}
\details{
This function is experimental and implements only a subset of the
variance estimators available in \code{fixest}. Several panel robust
estimators and user-supplied functions and covariance matrices are not
supported.
}
